Refactor: Isolate project headers from globals.h and standardize include order
- Optimized globals.h by removing major project headers (ntptimeclient.h, colordata.h, ledbuffer.h, ws281xgfx.h, socketserver.h, gfxbase.h).
- Enforced 'globals.h first' include rule project-wide.
- Resolved build failures across all environments (missing includes, static member order, palette linkage).
- Moved DecompressBuffer implementation to socketserver.cpp.
- Verified Arduino 3.0 compatibility for Network.h collisions.
