+++ include/effectmanager.h
@@ -35,12 +35,14 @@
 
 #pragma once
 
-#include <set>
+#include "globals.h"
 #include <algorithm>
 #include <functional>
 #include <math.h>
+#include <set>
 #include "effectfactories.h"
 #include "hub75gfx.h"
+#include "values.h"
 
 #define JSON_FORMAT_VERSION         1
 #define CURRENT_EFFECT_CONFIG_FILE  "/current.cfg"
