+++ include/effects/strip/paletteeffect.h
@@ -31,6 +31,7 @@
 #pragma once
 
 #include "effects.h"
+#include "values.h"
 
 template <typename TEffect>
 class PaletteEffectBase : public EffectWithId<TEffect>
