+++ src/debug_cli.cpp
@@ -25,26 +25,28 @@
 //    Implementation of debugging command line interface to NightDriver.
 //
 
+#include "globals.h"
+
 #include <algorithm>
-#include <charconv>
 #include <cctype>
+#include <charconv>
 #include <cstring>
 #include <esp_heap_caps.h>
 #include <esp_log.h>
-#include <esp_system.h> // esp_restart
+#include <esp_system.h>
 #include <freertos/FreeRTOS.h>
 #include <freertos/task.h>
+#include <FS.h>
+#include <optional>
+#include <SPIFFS.h>
 #include <string>
 #include <string_view>
 #include <vector>
-#include <optional>
 
-#include "globals.h"
+#include "colordata.h"
 #include "debug_cli.h"
 #include "effectmanager.h"
 #include "systemcontainer.h"
-#include <FS.h>
-#include <SPIFFS.h>
 
 namespace DebugCLI
 {
