+++ src/deviceconfig.cpp
@@ -28,9 +28,11 @@
 //
 //---------------------------------------------------------------------------
 
+#include "globals.h"
+
 #include <HTTPClient.h>
 #include <UrlEncode.h>
-#include "globals.h"
+
 #include "systemcontainer.h"
 
 extern const char timezones_start[] asm("_binary_config_timezones_json_start");
